requirements.txt
cvzone
opencv-python
mediapipe
torch
torchvision
pillow


